wip - Refactors calendar component structure and data flow
Improves component organization by introducing reusable components for calendar items and document details.

Streamlines data handling using a composable function to manage calendar data, filtering, and grouping.

Enhances notification display and data fetching with dedicated utilities.
